api route protected

+++ pages/api/news/[id].js
@@ -2,6 +2,8 @@ import news from "@/models/news";
 import dbConnect from "@/utils/database";
 import { promises as fs } from "fs";
 import formidable from "formidable";
+import { getSession } from "next-auth/react"
+
 
 dbConnect();
 
@@ -10,6 +12,7 @@ const requestModHandler = async (req, res) => {
     method,
     query: { id },
   } = req;
+  const session = await getSession({ req })
 
   switch (method) {
     case "GET":
@@ -24,80 +27,86 @@ const requestModHandler = async (req, res) => {
 
     case "POST":
       try {
-        const form = new formidable.IncomingForm({ keepExtensions: true });
 
-        const formParsePhotoSuccess = await new Promise((resolve, reject) => {
-          form.parse(req, async (err, fields, files) => {
-            if (err) {
-              reject(err);
-              return;
-            }
 
-            const newsInfo = {
-              fields,
-            };
+        if (session.user.isAdmin) {
+          const form = new formidable.IncomingForm({ keepExtensions: true });
 
-            const dateNow = new Date();
-
-            if (files.photoUpload) {
-              newsInfo.oldpath = files.photoUpload.filepath;
-              newsInfo.link = `/assets/images/news/${dateNow.getTime()}-${
-                files.photoUpload.originalFilename
-              }`;
-              newsInfo.newpath = `./public/assets/images/news/${dateNow.getTime()}-${
-                files.photoUpload.originalFilename
-              }`;
-            } else {
-              await news.updateOne(
-                { _id: id },
-                {
-                  date: dateNow,
-                  title: fields.title,
-                  description: fields.description,
-                }
-              );
-              return res
-                .status(200)
-                .json({ success: true, msg: "amjilttai edit hiile" });
-            }
-
-            resolve(newsInfo);
-          });
-        }).then(async (newsInfo) => {
-          let isSuccess = false;
-
-          await fs
-            .rename(newsInfo.oldpath, newsInfo.newpath, (err) => {
+          const formParsePhotoSuccess = await new Promise((resolve, reject) => {
+            form.parse(req, async (err, fields, files) => {
               if (err) {
-                console.log(err);
-                throw err;
+                reject(err);
+                return;
               }
-            })
-            .then(async () => {
-              isSuccess = true;
 
-              const myNews = {
-                date: new Date(),
-                title: newsInfo.fields.title,
-                description: newsInfo.fields.description,
-                photoLink: newsInfo.link,
+              const newsInfo = {
+                fields,
               };
 
-              await news.updateOne({ _id: id }, myNews);
+              const dateNow = new Date();
+
+              if (files.photoUpload) {
+                newsInfo.oldpath = files.photoUpload.filepath;
+                newsInfo.link = `/assets/images/news/${dateNow.getTime()}-${files.photoUpload.originalFilename
+                  }`;
+                newsInfo.newpath = `./public/assets/images/news/${dateNow.getTime()}-${files.photoUpload.originalFilename
+                  }`;
+              } else {
+                await news.updateOne(
+                  { _id: id },
+                  {
+                    date: dateNow,
+                    title: fields.title,
+                    description: fields.description,
+                  }
+                );
+                return res
+                  .status(200)
+                  .json({ success: true, msg: "amjilttai edit hiile" });
+              }
+
+              resolve(newsInfo);
             });
+          }).then(async (newsInfo) => {
+            let isSuccess = false;
 
-          return isSuccess;
-        });
+            await fs
+              .rename(newsInfo.oldpath, newsInfo.newpath, (err) => {
+                if (err) {
+                  console.log(err);
+                  throw err;
+                }
+              })
+              .then(async () => {
+                isSuccess = true;
 
-        if (formParsePhotoSuccess) {
-          return res
-            .status(200)
-            .json({ success: true, msg: "Successfully edited news" });
+                const myNews = {
+                  date: new Date(),
+                  title: newsInfo.fields.title,
+                  description: newsInfo.fields.description,
+                  photoLink: newsInfo.link,
+                };
+
+                await news.updateOne({ _id: id }, myNews);
+              });
+
+            return isSuccess;
+          });
+
+          if (formParsePhotoSuccess) {
+            return res
+              .status(200)
+              .json({ success: true, msg: "Successfully edited news" });
+          } else {
+            return res
+              .status(200)
+              .json({ success: false, msg: "Error editing news" });
+          }
         } else {
-          return res
-            .status(200)
-            .json({ success: false, msg: "Error editing news" });
+          return res.status(401).json({ success: false, msg: "You dont have a access" });
         }
+
+
       } catch (error) {
         console.log(error);
         res.status(400).json({ success: false });
@@ -106,8 +115,13 @@ const requestModHandler = async (req, res) => {
 
     case "DELETE":
       try {
-        await news.deleteOne({ _id: id });
-        res.status(200).json({ success: true, msg: "Successfully deleted" });
+        if (session.user.isAdmin) {
+          await news.deleteOne({ _id: id });
+          res.status(200).json({ success: true, msg: "Successfully deleted" });
+        } else {
+          return res.status(401).json({ success: false, msg: "You dont have a access" });
+        }
+
       } catch (error) {
         console.log(error);
         res.status(400).json({ success: false });
